requirements.txt
altgraph==0.16.1
appdirs==1.4.3
autoit==0.1
beautifulsoup4==4.7.1
berserker-resolver==2.0.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cmake==3.13.3
comtypes==1.1.7
cssselect==1.0.3
cycler==0.10.0
dlib==19.17.0
dnspython==1.15.0
dnspython3==1.15.0
docopt==0.6.2
ecdsa==0.13.3
esptool==2.8
fake-useragent==0.1.11
Flask==1.1.1
Flask-SocketIO==4.2.1
future==0.17.1
geoip2==2.9.0
idna==2.8
imutils==0.5.2
itsdangerous==1.1.0
Jinja2==2.10.3
keyboard==0.13.3
kiwisolver==1.1.0
lxml==4.3.3
macholib==1.11
MarkupSafe==1.1.1
matplotlib==3.1.0
maxminddb==1.5.1
mouse==0.7.1
MouseInfo==0.0.4
mysql-connector-python==8.0.18
numpy==1.16.2
offshoot==0.1.6
opencv-python==4.0.0.21
pandas==0.24.2
parse==1.11.1
pefile==2018.8.8
Pillow==6.0.0
protobuf==3.11.2
pyaes==1.6.1
PyAudio==0.2.11
PyAutoGUI==0.9.47
PyAutoIt==0.4
pyee==5.0.0
pygame==1.9.6
pygeoip==0.3.2
PyGetWindow==0.0.7
PyInstaller==3.4
PyMsgBox==1.0.7
PyMySQL==0.9.3
pynput==1.4.2
pyparsing==2.4.0
pyperclip==1.7.0
pyppeteer==0.0.25
PyQt4==4.11.4
PyQt5==5.11.3
PyQt5-sip==4.19.15
pyqt5-tools==5.11.3.1.4
pyqtgraph==0.10.0
pyquery==1.4.0
PyRect==0.1.4
PyScreeze==0.1.22
pyserial==3.4
pytesseract==0.3.2
python-dateutil==2.8.0
python-dotenv==0.10.1
python-engineio==3.11.1
python-socketio==4.4.0
pytraceroute==0.1.0
PyTweening==1.0.3
pytz==2018.9
pywin32==224
pywin32-ctypes==0.2.0
pywinauto==0.6.7
PyYAML==5.1.2
requests==2.21.0
requests-html==0.10.0
scapy==2.4.2
scipy==1.2.1
selenium==3.141.0
SimpleCV==1.3
six==1.12.0
soupsieve==1.9
tornado==6.0.3
tqdm==4.31.1
urllib3==1.24.1
w3lib==1.20.0
webb==0.9.2.5
websocket-client==0.56.0
websocket-server==0.4
websockets==7.0
Werkzeug==0.16.0
xlib==0.21
